tests/integration/requirements.txt
juju
requests
pytest-operator
python-keystoneclient
python-swiftclient
Pillow
pytest
pytest-cov